use single responsibility principle
The Host class shouldn't know how to send messages
or know any details about Telegram. This is just
bad design. The Host class should just know how to
track the state of the servers that are monitored.

This implicitly also fixes R0902 and R0913 that have
been nicely pointing out that code smell.